Move plugin meta data out of "plugin" package
The PluginMetaData class has to reside on the same level or higher as the input classes
for the content pack plugin discovery to work.

Refs Graylog2/graylog2-server#4909
